requirements.txt
easy_torch==1.3.2
easydict==1.13
einops==0.8.1
faiss==1.5.3
fastdtw==0.3.4
h5py==3.13.0
lib==4.0.0
matplotlib==3.10.1
mxnet==1.9.1
numpy==1.24.4
packaging==24.2
pandas==2.2.3
Pillow==11.1.0
positional_encodings==6.0.4
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setproctitle==1.3.4
sympy==1.13.1
timm==1.0.15
torch==2.6.0
torchdiffeq==0.2.5
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
tvm==1.0.0
umap==0.1.1
wandb==0.19.9
